requirements.txt
openai>=1.12.0
tiktoken>=0.6.0
streamlit>=1.32.0
python-dotenv>=1.0.0
rich>=13.7.0
pytest>=8.0.0
